Спринт 7. Тема 5 — Рефакторинг MVP. Урок 5/7 выполнен
