+++ MovieQuiz/Presentation/MovieQuizPresenter.swift
@@ -1,0 +1,32 @@
+//
+//  MovieQuizPresenter.swift
+//  MovieQuiz
+//
+//  Created by Сергей Розов on 30.04.2025.
+//
+
+import UIKit
+
+final class MovieQuizPresenter {
+    let questionsAmount = 10
+    private var currentQuestionIndex = 0
+
+    func isLastQuestion() -> Bool {
+        currentQuestionIndex == questionsAmount - 1
+    }
+    
+    func resetQuestionIndex() {
+        currentQuestionIndex = 0
+    }
+    
+    func switchToNextQuestion() {
+        currentQuestionIndex += 1
+    }
+    
+    func convert(model: QuizQuestionModel) -> QuizStepModel {
+        let image = UIImage(data: model.image) ?? UIImage()
+        let question = model.text
+        let questionNumber = "\(currentQuestionIndex + 1)/\(questionsAmount)"
+        return QuizStepModel(image: image, question: question, questionNumber: questionNumber)
+    }
+}
